move global func into the servlet
