+++ lib/msf/core/web_services/servlet/file_servlet.rb
@@ -2,7 +2,6 @@
 # Standard Library
 #
 
-require 'fileutils'
 module FileServlet
   def self.api_path
     '/api/v1/files'
@@ -48,7 +47,7 @@ module FileServlet
       sanitized_params = sanitize_params(params, env['rack.request.query_hash'])
       opts_path = sanitized_params[:path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
-      if safe_expand_path?(path) && File.exist?(path) && !File.directory?(path)
+      if get_db.safe_expand_path?(path) && File.exist?(path) && !File.directory?(path)
         send_file(path, buffer_size: 4096, stream: true)
       else
         result = { message: 'No such file' }
@@ -65,7 +64,7 @@ module FileServlet
         opts_path = sanitized_params[:path] || ''
         path = File.join(Msf::Config.rest_files_directory, opts_path)
         temp_path = sanitized_params[:file][:tempfile].path
-        if safe_expand_path?(path) && !File.exist?(path)
+        if get_db.safe_expand_path?(path) && !File.exist?(path)
           FileUtils.mkdir_p(File.dirname(path))
           FileUtils.cp_r(temp_path, path)
           result = { path: path }
@@ -86,7 +85,7 @@ module FileServlet
       opts = parse_json_request(request, true)
       opts_path = opts[:path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
-      if !rest_files_directory?(path) && !File.directory?(path) && safe_expand_path?(path)
+      if !get_db.rest_files_directory?(path) && !File.directory?(path) && get_db.safe_expand_path?(path)
         FileUtils.rm_rf(path)
         result = { path: path }
       else
@@ -104,8 +103,8 @@ module FileServlet
       opts_new_path = opts[:new_path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
       new_path = File.join(Msf::Config.rest_files_directory, opts_new_path)
-      if (!File.directory?(path) && safe_expand_path?(path)) && File.exist?(path) \
-        && (!File.directory?(new_path) && safe_expand_path?(new_path))
+      if (!File.directory?(path) && get_db.safe_expand_path?(path)) && File.exist?(path) \
+        && (!File.directory?(new_path) && get_db.safe_expand_path?(new_path))
         FileUtils.mv(path, new_path)
         result = { path: new_path }
       else
@@ -122,8 +121,8 @@ module FileServlet
       sanitized_params = sanitize_params(params, env['rack.request.query_hash'])
       opts_path = sanitized_params[:path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
-      if safe_expand_path?(path) && File.directory?(path)
-        result = list_local_path(path)
+      if get_db.safe_expand_path?(path) && File.directory?(path)
+        result = get_db.list_local_path(path)
       else
         result = { message: 'No such directory' }
       end
@@ -137,7 +136,7 @@ module FileServlet
       opts = parse_json_request(request, true)
       opts_path = opts[:path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
-      if safe_expand_path?(path)
+      if get_db.safe_expand_path?(path)
         FileUtils.mkdir_p(path)
         result = { path: path }
       else
@@ -153,7 +152,7 @@ module FileServlet
       opts = parse_json_request(request, true)
       opts_path = opts[:path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
-      if !rest_files_directory?(path) && safe_expand_path?(path) && File.directory?(path)
+      if !get_db.rest_files_directory?(path) && get_db.safe_expand_path?(path) && File.directory?(path)
         FileUtils.rm_rf(path)
         result = { path: path }
       else
@@ -171,8 +170,8 @@ module FileServlet
       opts_new_path = opts[:new_path] || ''
       path = File.join(Msf::Config.rest_files_directory, opts_path)
       new_path = File.join(Msf::Config.rest_files_directory, opts_new_path)
-      if (!rest_files_directory?(path) && safe_expand_path?(path) && File.directory?(path)) \
-        && (!rest_files_directory?(new_path) && safe_expand_path?(new_path))
+      if (!get_db.rest_files_directory?(path) && get_db.safe_expand_path?(path) && File.directory?(path)) \
+        && (!get_db.rest_files_directory?(new_path) && get_db.safe_expand_path?(new_path))
         FileUtils.mv(path, new_path)
         result = { path: new_path }
       else
@@ -202,7 +201,7 @@ module FileServlet
       search_file_paths = []
       utf8_buf = search_term.dup.force_encoding('UTF-8')
       # Return search keywords file path
-      if utf8_buf.valid_encoding? && safe_expand_path?(path) && File.directory?(path)
+      if utf8_buf.valid_encoding? && get_db.safe_expand_path?(path) && File.directory?(path)
         regex = Regexp.new(Regexp.escape(utf8_buf), true)
         Find.find(path) do |file_path|
           search_file_paths << file_path if (file_path =~ regex) && !File.directory?(file_path)
@@ -214,36 +213,4 @@ module FileServlet
       set_json_data_response(response: result)
     }
   end
-end
-
-def list_local_path(path)
-  # Enumerate each item...
-  tbl = []
-  files = Dir.entries(path)
-  files.each do |file|
-    file_path = File.join(path, file)
-    stat = File.stat(file_path)
-    row = {
-      name: file,
-      type: stat.ftype || '',
-      size: stat.size ? stat.size.to_s : '',
-      last_modified: stat.mtime || ''
-    }
-    next unless file != '.' && file != '..'
-
-    tbl << row
-  end
-  return tbl
-end
-
-def safe_expand_path?(path)
-  current_directory = File.expand_path(Msf::Config.rest_files_directory) + File::SEPARATOR
-  tested_path = File.expand_path(path) + File::SEPARATOR
-  tested_path.starts_with?(current_directory)
-end
-
-def rest_files_directory?(path)
-  tested_path = File.expand_path(path) + File::SEPARATOR
-  current_directory = File.expand_path(Msf::Config.rest_files_directory) + File::SEPARATOR
-  tested_path == current_directory
 end
